backend/requirements.txt
load-dotenv
fastapi
firebase_admin
bs4
langchain
langchain_chroma
langchain_community
langchain_openai